+++ packages/terra-functional-testing/tests/jest/commands/utils/index.test.js
@@ -5,9 +5,11 @@ const {
   dispatchCustomEvent,
   downloadScreenshots,
   getViewports,
+  getViewportSize,
   hideInputCaret,
   setApplicationLocale,
   setViewport,
+  setViewportSize,
 } = require('../../../../src/commands/utils');
 
 describe('index', () => {
@@ -18,8 +20,10 @@ describe('index', () => {
     expect(dispatchCustomEvent).toBeDefined();
     expect(downloadScreenshots).toBeDefined();
     expect(getViewports).toBeDefined();
+    expect(getViewportSize).toBeDefined();
     expect(hideInputCaret).toBeDefined();
     expect(setApplicationLocale).toBeDefined();
     expect(setViewport).toBeDefined();
+    expect(setViewportSize).toBeDefined();
   });
 });